chore: refactor packages naming
- add preview in com.zextras.mailbox.preview
- add usecase class/module
