+++ store/src/test/java/com/zextras/mailbox/preview/resource/PreviewQueryUtilTest.java
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-only
 
-package com.zextras.mailbox.resource.preview;
+package com.zextras.mailbox.preview.resource;
 
 import static org.junit.jupiter.api.Assertions.*;
