+++ store/src/test/resources/com/zextras/mailbox/preview/resource/attachment.txt
